feat: added contact anchor tag to link section
